Parsing Parameters was too Specific
Parsing parameters into integers is not valid for all CSI sequences.
Some parameters have leading '?' characters for example which cannot
result in an interpretation as an integer.  I've moved back to the
parser being more of a tokenizer.  Now parameters are passed to each
event handler as a slice of strings instead of integers so they can be
interpreted by the function as needed.  Probably the functions will have
to end up interpreting the strings back to integers in many cases, but
parsing appropriately before each function call seems like the wrong
place to do it.  Now each funciton can decide how to handle and type
it's parameters.